+++ wakamiti/wakamiti-api/src/main/java/es/iti/wakamiti/api/WakamitiAPI.java
@@ -3,8 +3,11 @@ package es.iti.wakamiti.api;
 import es.iti.commons.jext.ExtensionManager;
 import es.iti.wakamiti.api.plan.PlanSerializer;
 import es.iti.wakamiti.api.util.ResourceLoader;
+import imconfig.Configuration;
 
 
+import java.io.File;
+import java.nio.file.Path;
 import java.util.ServiceLoader;
 
 public interface WakamitiAPI {
@@ -16,7 +19,22 @@ public interface WakamitiAPI {
     WakamitiContributors contributors();
     ExtensionManager extensionManager();
     PlanSerializer planSerializer();
+
+
+    /**
+     * @deprecated Use either {@link #resourceLoader(File)} or {@link #resourceLoader(Configuration)}
+     * instead
+     */
+    @Deprecated
     ResourceLoader resourceLoader();
+
+    ResourceLoader resourceLoader(File workingDir);
+
+    ResourceLoader resourceLoader(Configuration configuration);
+
+
     void publishEvent(String eventType, Object data);
     String version();
+    Path workingDir(Configuration configuration);
+
 }